[Refactor] Move Sentry function to Sentry code
